requirements.txt
tweepy==3.10.0
pandas==1.3.3
nltk==3.6.3
python-dotenv==0.19.1
scikit-learn==0.24.2
